refactor(landing): show first agent showcase card as expanded by default until users hover on any of agent showcase cards.
- componentize agent showcase component
