+++ web-app/src/app/(landing)/components/agents-showcase/index.ts
@@ -1,0 +1,1 @@
+export { default as AgentsShowcase } from "./agents-showcase";
